Moved search bar and tag functionalities into new component
